refactor(cache): separate storage binding and key ownership
